refactor: split cli.py into a cli/ package
cli.py was 1 700 lines covering four entity types (proceedings, members,
bills, votes) plus the serve command. Break it into:

  src/concord/cli/
    __init__.py      root app, wiring, re-exports, main()
    _apps.py         the four stage Typer instances (no commands)
    _common.py       Progress helper, shared constants and parsers
    proceedings.py   scrape / load / index / run proceedings
    members.py       scrape / load / index / run members
    bills.py         scrape / scrape-enrich / load / index / run bills
    votes.py         scrape / load / index / run votes
    serve.py         serve command

_apps.py is the trick that avoids circular imports: entity modules import
the stage Typer() instances from there to register @scrape_app.command()
etc.; __init__.py imports the entity modules for their side effects, then
wires everything together. No cycles anywhere in the chain.

Public surface unchanged — concord.cli:main entry point, concord.cli.app,
concord.cli.ENV_OPENAI_API_KEY, concord.cli.MongoStorage all still
reachable from concord.cli. Two test lines updated: _run_pull/_run_load/
_run_index are now patched on concord.cli.proceedings (where they live),
and two cli_module.ENV_API_KEY references use the already-imported
ENV_API_KEY directly.
